Initialize Homebrew miniconda in globals
After a truly significant amount of frustration, friction, and research;
I have come to the conclusion that anaconda/miniconda should not be
managed by pyenv.

The ultimate reason for this is that Conda is not, strictly speaking, a
*Python* package manager, which is aparently a common misconception.

https://jakevdp.github.io/blog/2016/08/25/conda-myths-and-misconceptions

With this in mind, and after a nice tip from StackOverflow, I've decided
to abandon my original strategy of having pyenv-virtualenv manage all my
environments, and instead stick to a "seperation of concerns" strategy
wherein I use pyenv to manage python versions, regular venv or poetry to
manage general Python projects, and conda for data science projects.

ref: https://stackoverflow.com/a/58045893