Set VS Code terminal to login shell
A few days ago I discovered that I couldn't find any programs that are
sourced from my "/etc/path.d/" directory from inside the VS Code
terminal. This meant that I couldn't use Postgres.app's included CLI
commands from inside the integrated terminal, which was driving me
INSANE.

After some gooling and reading around, it turns out that this is because
the VS Code integrated terminal doesn't act as a login shell. Passing
"-l" to the terminal in settings forces this behavior, which allows for
my GPG and Postgres tooling to be seen.