:bug: Cleanup spaceship prompt auto-configuration
Spaceship prompt, while wonderful, doesn't yet have a flag that allows
you to opt out of its "helpful" setup behavior. Specifically, when the
package is finished installing it will automatically update your .zshrc
file to include the initialization commands.

I'm currently adding those myself in the lib/prompt file, so it's
causing my system to double initialize the prompt after install and
upgrade.

This is a quick and dirty fix using git restore, but it should work for
now.